Do not change logger on the import time (#71)
Summary:

Setting up loggers and in general executing logic with side-effect on the import time is a really bad idea. It messes with the logging setup of other projects even when this logic is not used. Logging should only be set up in the entry point, not on the module import time.

Reviewed By: krishnakumar-kapil

Differential Revision: D49770131