Do not change logger on the import time (#71)
Summary:
Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/71

Setting up loggers and in general executing logic with side-effect on the import time is a really bad idea. It messes with the logging setup of other projects even when this logic is not used. Logging should only be set up in the entry point, not on the module import time.

Reviewed By: fengxizhou, krishnakumar-kapil

Differential Revision: D49770131

fbshipit-source-id: 4156ed60d450fcd9245c4c15b68b77eab170d091