Add `include_last_profiler_step` option to include the last profiler step (#79)
Summary:
## What does this PR do?
Fixes https://github.com/facebookresearch/HolisticTraceAnalysis/issues/76 , see also https://github.com/facebookresearch/HolisticTraceAnalysis/issues/77 . This is a new PR initiated based on the branch from the cloned repo instead of from the forked repo.

When parsing the trace files, originally it was intentional that the last profiler step is not included because of some traces events in the last profiler step might be cut off during the trace collection. This PR adds `include_last_profiler_step` option to the `TraceAnalysis` class to allow users to switch on and off based on their needs.

## Before submitting

- [x] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
  - [ ] N/A
- [x] Did you write any new necessary tests?
  - [ ] N/A
- [ ] Did you make sure to update the docs?
  - [x] N/A
- [ ] Did you update the [changelog](https://github.com/facebookresearch/HolisticTraceAnalysis/blob/main/CHANGELOG.md)?
  - [x] N/A

Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/79

Reviewed By: leitian

Differential Revision: D50840133

Pulled By: anupambhatnagar

fbshipit-source-id: f5b47fedc512474879d765b538417b3ca8905b3b