Update critical path edge set computation (#92)
Summary:
## What does this PR do?
Fixes a bug in listing out the edges on the critical path.

### Details
Previously, after getting the ordered list of nodes on a critical path I was looking up all edges that had a src and dest in this list
```
for u, v in self.edges:
            e = self.edges[u, v]["object"]
            if u in critical_path_nodes_set and v in critical_path_nodes_set:   # <<< See here
                  self.critical_path_edges_set.add(e)
```

This is incorrect as pointed out by one of the users. We should only consider edges for nodes along the path itself.
For example, say the critical path node list is = [2, 6, 9, 10]. We should only include edges 2->6, 6->9 and 9->10. Previous logic was looking up any edges (u,v) with u,v belonging to [2, 6, 9, 10].

## Test
```
 pip install -e .;
 python3 -m unittest -v tests.test_critical_path_analysis.CriticalPathAnalysisTestCase
```

## Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
  - [x] N/A
- [x] Did you write any new necessary tests?
  - [ ] N/A
- [ ] Did you make sure to update the docs?
  - [x] N/A
- [x] Did you update the [changelog](https://github.com/facebookresearch/HolisticTraceAnalysis/blob/main/CHANGELOG.md)?
  - [ ] N/A

Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/92

Reviewed By: anupambhatnagar, fengxizhou

Differential Revision: D52424753

Pulled By: briancoutinho

fbshipit-source-id: abe15c1d701ac180c2dd91b8deaac82b1b3814ce