Fixing Unary Op evaluate issue (#112)
Summary:
Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/112

Without this diff, the critical_path_analysis on some traces fails with the error: `'UnaryOp' object has no attribute 'evaluate'`

This is apparently a pandas bug: https://github.com/pandas-dev/pandas/issues/16363, which has been solved a few times, but appears in different forms time, and again.

Reframing the query is solving the problem in critical_path_analysis for the time being.

Reviewed By: briancoutinho

Differential Revision: D54828775

fbshipit-source-id: 0bb52d1da7b5b143949c464b453955a666d31581