determine previous kernel for an event using the stream (#121)
Summary:
## What does this PR do?
<WIP>
Fixes https://github.com/facebookresearch/HolisticTraceAnalysis/issues/117
There are two fixes here.
### 1) Event Sync has stream -1
Support "Event Sync" event that actually has stream -1. This was changed for a large number to -1 in kineto but this caused the even sync to not work.

### 2) Cuda event record is stream aware.
For CUDA event synchronization we were previously considering the **previous kernel launched on the same thread**  before "cudaEventRecord()". This is actually incorrect as explained in https://github.com/facebookresearch/HolisticTraceAnalysis/issues/117. Ideally, we know which **stream an event was recorded on**. We should then mark the **previous kernel launched on that GPU and stream** as the one that will be source of a synchronization.

To fix behavior there is a change to the algorithm
1. Using "cuda_sync" events we can determine the CUDA event record -> cuda stream mapping.
2. Now, when searching backwards, instead of searching for previous kernel launch on a (pid, tid), we search back on a (gpu, stream). This ensure we are accurate in finding the kernel that ran on the stream before the event was recorded.

## Testing

Here is a simple reproduction of the issue = [full code](https://gist.github.com/briancoutinho/3556b66f407cd16322a7b6968429025f)
 An outline of what the code does
 ```
 # run kernel 1 on stream1 (stream = 20)
 with torch.cuda.stream(stream1):
     result = torch.matmul(matrix_a, matrix_b)

# run kernel 2 on stream3 (stream = 28)
with torch.cuda.stream(stream3):
    result = torch.matmul(matrix_a, matrix_b)

# record an event on stream1 / stream = 20.
event.record(stream1)
# make kernels on stream 2 wait on the above event.
event.wait(stream2)

# run a kernel3  on stream 2
with torch.cuda.stream(stream2):
     result1 = torch.matmul(matrix_c, result)
 ```

When we run the HTA Critical path analysis since we only look for the previous kernel launch before event record (after kernel 2 is launched) we are incorrectly adding dependency between the kernel 2 -> kernel 3

<img width="1119" alt="cp_before_fix" src="https://github.com/facebookresearch/HolisticTraceAnalysis/assets/6922212/747359e2-15fd-471c-b857-6326ac8204fc">

Previously, the cude event record dataframe provides a dependency on kernel 2 (correlation 57)
<img width="1141" alt="hta_actually_before_df" src="https://github.com/facebookresearch/HolisticTraceAnalysis/assets/6922212/aea39222-cd30-476d-8221-333731559624">

After the fix the new code has correct dependency to kernel 1 correlation 27.
<img width="902" alt="hta_before_df" src="https://github.com/facebookresearch/HolisticTraceAnalysis/assets/6922212/70d027bf-49c4-49b2-b272-6f91a673b305">

And the dependency is marked correctly
<img width="1096" alt="hta_after_final" src="https://github.com/facebookresearch/HolisticTraceAnalysis/assets/6922212/784b5a71-ca7f-47d3-bfa8-ba03f3e6795e">

## TODO
* Add unit test from collected trace
* Update docs

## Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
  - [ ] N/A
- [ ] Did you write any new necessary tests?
  - [ ] N/A
- [ ] Did you make sure to update the docs?
  - [ ] N/A
- [ ] Did you update the [changelog](https://github.com/facebookresearch/HolisticTraceAnalysis/blob/main/CHANGELOG.md)?
  - [ ] N/A

Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/121

Reviewed By: gajjanag

Differential Revision: D56002659

Pulled By: briancoutinho

fbshipit-source-id: ac04df5f0d22dede2dc33f4119f78850bf67b690