0507 add metadata parser ijson (#132)
Summary:
## What does this PR do?
This PR adds a final missing feature in ijson parser. We need to obtain trace metadata in addition to the events.
This change uses ijson.parse() API to construct the trace metadata and immediately stops when it sees "traceEvents"
There is more inline documentation for it.

## Unit Testing
Added a new test
  `pytest tests/test_trace_parse.py -k test_ijson_metadata_reader`

And metadata tests on main trace parser class
  `pytest tests/test_trace_parse.py -k test_trace_metadata`

## Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
  - [x] N/A
- [x] Did you write any new necessary tests?
  - [ ] N/A
- [ ] Did you make sure to update the docs?
  - [x] N/A
- [] Did you update the [changelog](https://github.com/facebookresearch/HolisticTraceAnalysis/blob/main/CHANGELOG.md)?
  - [x] N/A

Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/132

Reviewed By: jj10306

Differential Revision: D57296015

Pulled By: briancoutinho

fbshipit-source-id: 50aec92908b796207ac188915e3b056a349b0297