Add required columns in parser config
Summary:
Context is in task. We've been seeing some cases when trace collected is empty and that causes some parsing error

**This diff**
defines a default list of columns that are required in order for trace df to be considered not empty. In case of empty, we return a df with returned columns to avoid following steps triggering additional column not found errors

**Incoming:**
Handle empty trace case from caller

Reviewed By: fengxizhou

Differential Revision: D66393379