Add required columns in parser config (#207)
Summary:
Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/207

Context is in task. We've been seeing some cases when trace collected is empty and that causes some parsing error

**This diff**
defines a default list of columns that are required in order for trace df to be considered not empty. In case of empty, we return a df with returned columns to avoid following steps triggering additional column not found errors

**Incoming:**
Handle empty trace case from caller

Reviewed By: fengxizhou

Differential Revision: D66393379

fbshipit-source-id: 70d21df6ce26baff43f176225e69902c90fcd420