fix kernel lauch delay
Summary:
This diff fixes the kernel launch statistics for HTA.

When we are calculating the launch delay (assuming the delay is the time from the end of kernel launch to the start of the kernel execution. the equation should be `(gpu_kernel_start_ts) - (cpu_kernel_start_ts + duration_cpu)`? The sign for `cpu_duration` is flipped here https://fburl.com/code/tckgei9p, which is incorrect. This fix fixes that.

Moreover, we clip the kernel launch delay to 0 is the value is negative i.e kernel execution starts before the launch has finished.

Differential Revision: D68877268