fix kernel launch delay (#214)
Summary:
Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/214

This diff fixes the kernel launch statistics for HTA.

When we are calculating the launch delay (assuming the delay is the time from the end of kernel launch to the start of the kernel execution. the equation should be `(gpu_kernel_start_ts) - (cpu_kernel_start_ts + duration_cpu)`? The sign for `cpu_duration` is flipped here https://fburl.com/code/tckgei9p, which is incorrect. This fix fixes that.

Moreover, we clip the kernel launch delay to 0 is the value is negative i.e kernel execution starts before the launch has finished.

Reviewed By: fengxizhou

Differential Revision: D68877268

fbshipit-source-id: 7eea89e270c2a7a6422af059aecb8e7604cf1ff1