: Filter out GPU events with no correlation
Summary:
The inner merge seems to be creating all possible combinations of correlation (-1, -1) from both cpu and gpu data frames and OOM'ing.

This diff fixes the issue by filtering out GPU events with index correlation < 0

Note: This wasnt an issue before because this seems to happen (on nvidia traces) only when stream == 0. The change made to enable AMD caused this corner case to trigger.

Differential Revision: D72218256