Filter out GPU events with no correlation (#237)
Summary:
Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/237

The inner merge seems to be creating all possible combinations of correlation (-1, -1) from both cpu and gpu data frames and OOM'ing.

This diff fixes the issue by filtering out GPU events with index correlation < 0

Note: This wasnt an issue before because this seems to happen (on nvidia traces) only when stream == 0. The change made to enable AMD caused this corner case to trigger.

Reviewed By: briancoutinho

Differential Revision: D72218256

fbshipit-source-id: 4dd8930029c6292151fd2eded81d796f3d9d155e