make op kernels and delay generalized
Summary: `get_aten_op_kernels_and_delay`  was added in D73267800 that allowed obtianing ssociated kernels related to Aten Op and Aten Op overhead. We wish to make it generalized for other ops as well. So, we let the users specify op_name_pattern and default to "aten::" when none is provided

Differential Revision: D73855981