changes to support for Hardware Het. effort part 1
Summary:
A collection of changes that we need to measure MTIA traces. Most of this work is mainly catch up of open/HTA with work from fengxizhou on p9e/HTA with minor modifications to adhere to the structure within open/HTA.

We also bring in and add tests for these changes.

Differential Revision: D75942321