Add end-to-end testing mechanism. (#269)
Summary:
Pull Request resolved: https://github.com/facebookresearch/HolisticTraceAnalysis/pull/269

Allows the developer to run end-to-end tests by providing an input trace and expected output. Usage is described in the code comment.

The goal is to allow me to assert meta traces produce expected output. Since we don't want to push meta traces to the open source repository, we're going to instead just keep an internal copy of test cases we care about, and copy them over whenever we want to change.

For example, D77951398 adds some functionality to merge forward/backward passes to the open-source repo. But the test case uses real meta traces, so it'll be kept separately in D78127548.

Reviewed By: seansundor

Differential Revision: D78124539