Merge backward pass into main thread (forward pass)
Summary:
Merges backward pass into main thread (where possible).

For training jobs, backward pass is usually run on a separate thread. At the moment, CPA only considers one thread at a time, so the backward pass thread is usually omitted from CPA.

But backward pass events and forward pass events should have no overlap, so it should be safe to just merge the backward pass into the forward pass.

In this change, we try to detect the backward pass thread as the only thread containing events with "autograd" in the name. If there is exactly one such thread, we assume that is the backward pass thread, and merge its events into the main thread (by simply changing their tid to the main tid when constructing the CallStack).

Reviewed By: seansundor

Differential Revision: D77951398