Support data loading operations.
Summary:
Supports including data loading events on the critical path. In the open-source library, we allow passing in a list of regexes to consider as data loading. For our internal use case, we pass in ["data_loading" and "copy_batch_to_gpu"].

These are both user annotations, so previously they were completely omitted from the critical path. But we've seen that often there are no specific cpu or gpu ops nested inside, and that these are actually the best leaf operators to use.

Note: I'll follow up this diff with another one that adds a new "data loading" category in the CPA breakdown.

Reviewed By: seansundor

Differential Revision: D78374879