Add threshold for kernel-kernel delays
Summary:
Only consider a possible edge as a kernel-kernel delay if the previous event occurred within a threshold (1.5 ms).

Typically, queued consecutive kernels execute within < 10 microseconds of the previous kernel executing. Our code did not previously take this into account; instead, it identified kernel-kernel delays as any consecutive kernels where there are kernels in the queue (queue_length > 1).

But there are cases where there can be kernels in the queue BUT there is actually some other event blocking progress. For example, consider the following:

CPU stream:
- launches kernels A, B, C
- issues an all-to-all collective
- launches kernels D, E, F (before all-to-all executes)

GPU stream:
- executes kernels A, B, C
- executes all-to-all
- executes kernel D, E, F

Prior to this change, when the kernel D gets executed, it'll be the case that there are other kernels queued up (E, F), so queue_length > 1. Because of this, it'll assign an edge with a very long time duration from (C)->(D) as a "kernel-kernel" delay.

In reality, C gets executed, then the all-to-all gets executed, then D gets executed. There should be no kernel-kernel delay edge between C and D.

Reviewed By: seansundor

Differential Revision: D78834848