Adjust code to use the new singletrace module.
The approach is to get the trace DataFrame directly from
TraceCollection object instead of getting Trace object first
and then accessing its DataFrame attribute.

My current approach is to use the Trace objects directly when:
- A method has more than one argument that is aggregated by Trace
  object. This way we reduce the number of parameters in that method.
- A method passes the whole Trace object to another method.
- A method returns more than one attribute that is aggregated by Trace
  object. This way we reduce the number of return values.

With this approach we don't need to modify all the methods that use
pd.DataFrame as input/output parameters.
When there is a need to pass the Trace object to use its methods or
attributes, we can simply adjust the methods that need it.

Signed-off-by: Benedykt Bela <benedykt.bela@intel.com>