⬆️ Resolve networkx python package 3.6 version issue. (#16)
* Resolve networkx python package 3.6 version issue.

HTA has the requirement that networkx package should be greater than or
equal to 3.1 version. For some reason, the version 3.6 of networkx
package causes HTA to fails during critical path analysis.
What is interesting, the version 3.6.1 of networkx package works fine
with HTA. Therefore, to avoid the installation of the problematic
version 3.6, we exclude it in the requirements.txt file.
Issue in networkx repository: https://github.com/networkx/networkx/issues/8367

Signed-off-by: Benedykt Bela <benedykt.bela@intel.com>

* Change comment.

Signed-off-by: Benedykt Bela <benedykt.bela@intel.com>

---------

Signed-off-by: Benedykt Bela <benedykt.bela@intel.com>